+++ src/components/discover-sheet/UniswapPoolsSection.js
@@ -26,7 +26,13 @@ const DefaultShowMoreButton = ({ backgroundColor, color, onPress }) => (
         paddingHorizontal={12}
         paddingTop={android ? 3 : 7}
       >
-        <Text align="center" color={color} size="lmedium" weight="heavy">
+        <Text
+          align="center"
+          color={color}
+          lineHeight={30}
+          size="lmedium"
+          weight="heavy"
+        >
           Show more
         </Text>
       </Row>
@@ -64,10 +70,10 @@ const PoolListButton = styled(ButtonPressAnimation).attrs({
         border-radius: 12px;
         height: 30px;
         padding-horizontal: 8px;
-        padding-top: ${ios ? 6 : 4}px;
+        padding-top: ${ios ? 6.5 : 4.5}px;
       `
       : `
-        padding-top: ${ios ? 6 : 4}px;
+        padding-top: ${ios ? 6.5 : 4.5}px;
       `}
 `;