+++ src/redux/settings.js
@@ -1,20 +1,19 @@
 import analytics from '@segment/analytics-react-native';
-import { NativeCurrencyKeys } from '../entities/nativeCurrencyTypes';
+import { updateLanguage } from '../languages';
+import { NativeCurrencyKeys } from '@rainbow-me/entities';
 import {
   getNativeCurrency,
   getNetwork,
   saveLanguage,
   saveNativeCurrency,
   saveNetwork,
-} from '../handlers/localstorage/globalSettings';
-import { web3SetHttpProvider } from '../handlers/web3';
-import networkTypes from '../helpers/networkTypes';
-import { updateLanguage } from '../languages';
+} from '@rainbow-me/handlers/localstorage/globalSettings';
+import { web3SetHttpProvider } from '@rainbow-me/handlers/web3';
+import networkTypes from '@rainbow-me/helpers/networkTypes';
+import { dataResetState } from '@rainbow-me/redux/data';
+import { explorerClearState, explorerInit } from '@rainbow-me/redux/explorer';
 import { saveString } from '../model/keychain';
-
-import { ethereumUtils } from '../utils';
-import { dataResetState } from './data';
-import { explorerClearState, explorerInit } from './explorer';
+import { ethereumUtils } from '@rainbow-me/utils';
 import logger from 'logger';
 
 // -- Constants ------------------------------------------------------------- //